requirements.txt
PyYAML
python-novaclient